Final Fixes/Adjustments for Blackstone
More bug fixing, more map fixes and adjustments. Made some changes as indicated on the PR comments. Simple animals will take fire damage now.

Some changes to the crew tracking computer to allow it to better track squads as squad colors were limited; now it has room to add additional ones for proper coloring on the tracker.

Made most of the terrain in Blackstone destructible, added some procs and defines to work with that.
Refactored vent crawling/crawling animations a little bit and enabled ambush spawns for ladders (including hatch ladders that I introduced).

A lot of work on GLOB.data_core and the crew manifest. Refactored all of the manifest compiling, introducecd defines to better manage it, and added a few checks for data.

All three data computers (security, medical, skills), were slightly altered to allow faction-based checking. It's not great, but the alternative is to rewrite how data_core works entirely. Though that wouldn't be a bad idea given how dated it is, and records in general, that's out of the scope of this PR.

+++ code/game/machinery/computer/medical.dm
@@ -18,6 +18,7 @@
 	var/a_id = null
 	var/temp = null
 	var/printing = null
+	var/factions = list(FACTION_MARINE)
 
 /obj/structure/machinery/computer/med_data/verb/eject_id()
 	set category = "Object"
@@ -73,7 +74,8 @@
 					dat += "<B>Record List</B>:<HR>"
 					if(!isnull(GLOB.data_core.general))
 						for(var/datum/data/record/R in sortRecord(GLOB.data_core.general))
-							dat += text("<A href='?src=\ref[];d_rec=\ref[]'>[]: []<BR>", src, R, R.fields["id"], R.fields["name"])
+							if(R.fields["mob_faction"] in factions) //Only for the faction(s) we want.
+								dat += text("<A href='?src=\ref[];d_rec=\ref[]'>[]: []<BR>", src, R, R.fields["id"], R.fields["name"])
 							//Foreach goto(132)
 					dat += text("<HR><A href='?src=\ref[];screen=1'>Back</A>", src)
 				if(3.0)
@@ -213,10 +215,13 @@
 				src.temp = text("Are you sure you wish to delete all records?<br>\n\t<A href='?src=\ref[];temp=1;del_all2=1'>Yes</A><br>\n\t<A href='?src=\ref[];temp=1'>No</A><br>", src, src)
 
 			if (href_list["del_all2"])
-				for(var/datum/data/record/R as anything in GLOB.data_core.medical)
-					GLOB.data_core.medical -= R
-					qdel(R)
-					//Foreach goto(494)
+				for(var/datum/data/record/R in GLOB.data_core.general)
+					if(R.fields["mob_faction"] in factions) //Only for the faction(s) we want.
+						for(var/datum/data/record/E as anything in GLOB.data_core.medical)
+							if(E.fields["name"] == R.fields["name"] && E.fields["id"] == R.fields["id"])
+								GLOB.data_core.security -= E
+								qdel(E)
+
 				src.temp = "All records deleted."
 
 			if (href_list["field"])
@@ -450,7 +455,7 @@
 					else
 						P.info += "<B>Medical Record Lost!</B><BR>"
 					P.info += "</TT>"
-					P.info += text("<BR><HR><font size = \"1\"><I>This report was printed by [] [].<BR>The [MAIN_SHIP_NAME],[]/[], []</I></font><BR>\n<span class=\"paper_field\"></span>",last_user_rank,last_user_name,time2text(world.timeofday, "MM/DD"),game_year,worldtime2text())
+					P.info += text("<BR><HR><font size = \"1\"><I>This report was printed by [] [].<BR>The governing medical body,[]/[], []</I></font><BR>\n<span class=\"paper_field\"></span>",last_user_rank,last_user_name,time2text(world.timeofday, "MM/DD"),game_year,worldtime2text())
 					src.printing = null
 
 			if(href_list["print_bs"])//Prints latest body scan
@@ -473,7 +478,7 @@
 								break
 							else
 								P.info += "No scan on record."
-					P.info += text("<BR><HR><font size = \"1\"><I>This report was printed by [] [].<BR>The [MAIN_SHIP_NAME],  []/[], []</I></font><BR>\n<span class=\"paper_field\"></span>",last_user_rank,last_user_name,time2text(world.timeofday, "MM/DD"),game_year,worldtime2text())
+					P.info += text("<BR><HR><font size = \"1\"><I>This report was printed by [] [].<BR>The governing medical body,  []/[], []</I></font><BR>\n<span class=\"paper_field\"></span>",last_user_rank,last_user_name,time2text(world.timeofday, "MM/DD"),game_year,worldtime2text())
 					src.printing = null
 
 
@@ -487,27 +492,30 @@
 	if(inoperable())
 		return
 
-	for(var/datum/data/record/R as anything in GLOB.data_core.medical)
-		if(prob(10/severity))
-			switch(rand(1,6))
-				if(1)
-					R.fields["name"] = "[pick(pick(first_names_male), pick(first_names_female))] [pick(last_names)]"
-				if(2)
-					R.fields["sex"] = pick("Male", "Female")
-				if(3)
-					R.fields["age"] = rand(5, 85)
-				if(4)
-					R.fields["b_type"] = pick("A-", "B-", "AB-", "O-", "A+", "B+", "AB+", "O+")
-				if(5)
-					R.fields["p_stat"] = pick("*SSD*", "Active", "Physically Unfit", "Disabled")
-				if(6)
-					R.fields["m_stat"] = pick("*Insane*", "*Unstable*", "*Watch*", "Stable")
-			continue
+	for(var/datum/data/record/E in GLOB.data_core.general)
+		if(E.fields["mob_faction"] in factions)
+			for(var/datum/data/record/R as anything in GLOB.data_core.medical)
+				if (E.fields["name"] == R.fields["name"] && E.fields["id"] == R.fields["id"])
+					if(prob(10/severity))
+						switch(rand(1,6))
+							if(1)
+								R.fields["name"] = "[pick(pick(first_names_male), pick(first_names_female))] [pick(last_names)]"
+							if(2)
+								R.fields["sex"] = pick("Male", "Female")
+							if(3)
+								R.fields["age"] = rand(5, 85)
+							if(4)
+								R.fields["b_type"] = pick("A-", "B-", "AB-", "O-", "A+", "B+", "AB+", "O+")
+							if(5)
+								R.fields["p_stat"] = pick("*SSD*", "Active", "Physically Unfit", "Disabled")
+							if(6)
+								R.fields["m_stat"] = pick("*Insane*", "*Unstable*", "*Watch*", "Stable")
+						continue
 
-		else if(prob(1))
-			GLOB.data_core.medical -= R
-			qdel(R)
-			continue
+					else if(prob(1))
+						GLOB.data_core.medical -= R
+						qdel(R)
+						continue
 
 
 /obj/structure/machinery/computer/med_data/laptop
@@ -515,3 +523,9 @@
 	desc = "Cheap Weyland-Yutani Laptop."
 	icon_state = "medlaptop"
 	density = FALSE
+
+/obj/structure/machinery/computer/med_data/uscm_ground
+	icon_state = "medlaptop"
+	req_one_access = null
+	req_access = list(ACCESS_USCM_GROUND_MEDICAL)
+	factions = list(FACTION_USCM_GROUND)
