+++ code/game/machinery/computer/skills.dm
@@ -20,7 +20,7 @@
 	//Sorting Variables
 	var/sortBy = "name"
 	var/order = 1 // -1 = Descending - 1 = Ascending
-
+	var/factions = list(FACTION_MARINE)
 
 /obj/structure/machinery/computer/skills/attackby(obj/item/O as obj, user as mob)
 	if(istype(O, /obj/item/card/id) && !scan)
@@ -37,9 +37,7 @@
 /obj/structure/machinery/computer/skills/attack_hand(mob/user as mob)
 	if(..())
 		return
-	if (src.z > 6)
-		to_chat(user, SPAN_DANGER("<b>Unable to establish a connection</b>: \black You're too far away from the station!"))
-		return
+
 	var/dat
 
 	if (temp)
@@ -67,13 +65,16 @@
 <th><A href='?src=\ref[src];choice=Sorting;sort=rank'>Rank</A></th>
 <th><A href='?src=\ref[src];choice=Sorting;sort=fingerprint'>Fingerprints</A></th>
 </tr>"}
+
 					if(!isnull(GLOB.data_core.general))
 						for(var/datum/data/record/R in sortRecord(GLOB.data_core.general, sortBy, order))
-							for(var/datum/data/record/E in GLOB.data_core.security)
-							dat += "<tr><td><A href='?src=\ref[src];choice=Browse Record;d_rec=\ref[R]'>[R.fields["name"]]</a></td>"
-							dat += "<td>[R.fields["id"]]</td>"
-							dat += "<td>[R.fields["rank"]]</td>"
-						dat += "</table><hr width='75%' />"
+							if(R.fields["mob_faction"] in factions) //Only for the faction(s) we want.
+								for(var/datum/data/record/E in GLOB.data_core.security)
+									dat += "<tr><td><A href='?src=\ref[src];choice=Browse Record;d_rec=\ref[R]'>[R.fields["name"]]</a></td>"
+									dat += "<td>[R.fields["id"]]</td>"
+									dat += "<td>[R.fields["rank"]]</td>"
+								dat += "</table><hr width='75%' />"
+
 					dat += "<A href='?src=\ref[src];choice=Record Maintenance'>Record Maintenance</A><br><br>"
 					dat += "<A href='?src=\ref[src];choice=Log Out'>{Log Out}</A>"
 				if(2.0)
@@ -261,9 +262,12 @@ What a mess.*/
 				temp += "<a href='?src=\ref[src];choice=Clear Screen'>No</a>"
 
 			if ("Purge All Records")
-				for(var/datum/data/record/R in GLOB.data_core.security)
-					GLOB.data_core.security -= R
-					qdel(R)
+				for(var/datum/data/record/R in GLOB.data_core.general)
+					if(R.fields["mob_faction"] in factions) //Only for the faction(s) we want.
+						for(var/datum/data/record/E in GLOB.data_core.security)
+							if(E.fields["name"] == R.fields["name"] && E.fields["id"] == R.fields["id"])
+								GLOB.data_core.security -= E
+								qdel(E)
 				temp = "All Employment records deleted."
 
 			if ("Delete Record (ALL)")
@@ -273,7 +277,7 @@ What a mess.*/
 					temp += "<a href='?src=\ref[src];choice=Clear Screen'>No</a>"
 //RECORD CREATE
 			if ("New Record (General)")
-				active1 = CreateGeneralRecord()
+				active1 = CreateGeneralRecord(factions[1])
 
 //FIELD FUNCTIONS
 			if ("Edit Field")
@@ -351,25 +355,27 @@ What a mess.*/
 	if(inoperable())
 		return
 
-	for(var/datum/data/record/R in GLOB.data_core.security)
-		if(prob(10/severity))
-			switch(rand(1,6))
-				if(1)
-					R.fields["name"] = "[pick(pick(first_names_male), pick(first_names_female))] [pick(last_names)]"
-				if(2)
-					R.fields["sex"] = pick("Male", "Female")
-				if(3)
-					R.fields["age"] = rand(5, 85)
-				if(4)
-					R.fields["criminal"] = pick("None", "*Arrest*", "Incarcerated", "Released")
-				if(5)
-					R.fields["p_stat"] = pick("*Unconcious*", "Active", "Physically Unfit")
-				if(6)
-					R.fields["m_stat"] = pick("*Insane*", "*Unstable*", "*Watch*", "Stable")
-			continue
+	for(var/datum/data/record/E in GLOB.data_core.general)
+		if(E.fields["mob_faction"] in factions)
+			for(var/datum/data/record/R in GLOB.data_core.security)
+				if(prob(10/severity))
+					switch(rand(1,6))
+						if(1)
+							R.fields["name"] = "[pick(pick(first_names_male), pick(first_names_female))] [pick(last_names)]"
+						if(2)
+							R.fields["sex"] = pick("Male", "Female")
+						if(3)
+							R.fields["age"] = rand(5, 85)
+						if(4)
+							R.fields["criminal"] = pick("None", "*Arrest*", "Incarcerated", "Released")
+						if(5)
+							R.fields["p_stat"] = pick("*Unconcious*", "Active", "Physically Unfit")
+						if(6)
+							R.fields["m_stat"] = pick("*Insane*", "*Unstable*", "*Watch*", "Stable")
+					continue
 
-		else if(prob(1))
-			GLOB.data_core.security -= R
-			qdel(R)
-			continue
+				else if(prob(1))
+					GLOB.data_core.security -= R
+					qdel(R)
+					continue
